sensor-hub/requirements.txt
bleak>=0.21.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Data transmission
python-socketio[client]>=5.10.0
requests>=2.31.0
